requirements.txt
llama-index
pypdf
llama-index-llms-huggingface
llama-index-embeddings-huggingface
faiss-cpu
llama-index-vector-stores-faiss
groq
